refactor(router): change `Recognizer`s -> `PathRoute` and associated changes
