+++ modules/angular2/test/router/param_route_path_spec.ts
@@ -10,92 +10,92 @@ import {
   SpyObject
 } from 'angular2/testing_internal';
 
-import {PathRecognizer} from 'angular2/src/router/rules/path_recognizer';
+import {ParamRoutePath} from 'angular2/src/router/rules/param_route_path';
 import {parser, Url} from 'angular2/src/router/rules/url_parser';
 
 export function main() {
   describe('PathRecognizer', () => {
 
     it('should throw when given an invalid path', () => {
-      expect(() => new PathRecognizer('/hi#'))
+      expect(() => new ParamRoutePath('/hi#'))
           .toThrowError(`Path "/hi#" should not include "#". Use "HashLocationStrategy" instead.`);
-      expect(() => new PathRecognizer('hi?'))
+      expect(() => new ParamRoutePath('hi?'))
           .toThrowError(`Path "hi?" contains "?" which is not allowed in a route config.`);
-      expect(() => new PathRecognizer('hi;'))
+      expect(() => new ParamRoutePath('hi;'))
           .toThrowError(`Path "hi;" contains ";" which is not allowed in a route config.`);
-      expect(() => new PathRecognizer('hi='))
+      expect(() => new ParamRoutePath('hi='))
           .toThrowError(`Path "hi=" contains "=" which is not allowed in a route config.`);
-      expect(() => new PathRecognizer('hi('))
+      expect(() => new ParamRoutePath('hi('))
           .toThrowError(`Path "hi(" contains "(" which is not allowed in a route config.`);
-      expect(() => new PathRecognizer('hi)'))
+      expect(() => new ParamRoutePath('hi)'))
           .toThrowError(`Path "hi)" contains ")" which is not allowed in a route config.`);
-      expect(() => new PathRecognizer('hi//there'))
+      expect(() => new ParamRoutePath('hi//there'))
           .toThrowError(`Path "hi//there" contains "//" which is not allowed in a route config.`);
     });
 
     describe('querystring params', () => {
       it('should parse querystring params so long as the recognizer is a root', () => {
-        var rec = new PathRecognizer('/hello/there');
+        var rec = new ParamRoutePath('/hello/there');
         var url = parser.parse('/hello/there?name=igor');
-        var match = rec.recognize(url);
+        var match = rec.matchUrl(url);
         expect(match.allParams).toEqual({'name': 'igor'});
       });
 
       it('should return a combined map of parameters with the param expected in the URL path',
          () => {
-           var rec = new PathRecognizer('/hello/:name');
+           var rec = new ParamRoutePath('/hello/:name');
            var url = parser.parse('/hello/paul?topic=success');
-           var match = rec.recognize(url);
+           var match = rec.matchUrl(url);
            expect(match.allParams).toEqual({'name': 'paul', 'topic': 'success'});
          });
     });
 
     describe('matrix params', () => {
       it('should be parsed along with dynamic paths', () => {
-        var rec = new PathRecognizer('/hello/:id');
+        var rec = new ParamRoutePath('/hello/:id');
         var url = new Url('hello', new Url('matias', null, null, {'key': 'value'}));
-        var match = rec.recognize(url);
+        var match = rec.matchUrl(url);
         expect(match.allParams).toEqual({'id': 'matias', 'key': 'value'});
       });
 
       it('should be parsed on a static path', () => {
-        var rec = new PathRecognizer('/person');
+        var rec = new ParamRoutePath('/person');
         var url = new Url('person', null, null, {'name': 'dave'});
-        var match = rec.recognize(url);
+        var match = rec.matchUrl(url);
         expect(match.allParams).toEqual({'name': 'dave'});
       });
 
       it('should be ignored on a wildcard segment', () => {
-        var rec = new PathRecognizer('/wild/*everything');
+        var rec = new ParamRoutePath('/wild/*everything');
         var url = parser.parse('/wild/super;variable=value');
-        var match = rec.recognize(url);
+        var match = rec.matchUrl(url);
         expect(match.allParams).toEqual({'everything': 'super;variable=value'});
       });
 
       it('should set matrix param values to true when no value is present', () => {
-        var rec = new PathRecognizer('/path');
+        var rec = new ParamRoutePath('/path');
         var url = new Url('path', null, null, {'one': true, 'two': true, 'three': '3'});
-        var match = rec.recognize(url);
+        var match = rec.matchUrl(url);
         expect(match.allParams).toEqual({'one': true, 'two': true, 'three': '3'});
       });
 
       it('should be parsed on the final segment of the path', () => {
-        var rec = new PathRecognizer('/one/two/three');
+        var rec = new ParamRoutePath('/one/two/three');
 
         var three = new Url('three', null, null, {'c': '3'});
         var two = new Url('two', three, null, {'b': '2'});
         var one = new Url('one', two, null, {'a': '1'});
 
-        var match = rec.recognize(one);
+        var match = rec.matchUrl(one);
         expect(match.allParams).toEqual({'c': '3'});
       });
     });
 
     describe('wildcard segment', () => {
       it('should return a url path which matches the original url path', () => {
-        var rec = new PathRecognizer('/wild/*everything');
+        var rec = new ParamRoutePath('/wild/*everything');
         var url = parser.parse('/wild/super;variable=value/anotherPartAfterSlash');
-        var match = rec.recognize(url);
+        var match = rec.matchUrl(url);
         expect(match['urlPath']).toEqual('wild/super;variable=value/anotherPartAfterSlash');
       });
     });
